switching discriminator to non trainable

+++ tools/visualise.py
@@ -6,7 +6,7 @@ import datetime
 
 
 def show_images(imgs, title="", save_instead=False):
-    imgs = (imgs*255).astype(int)
+    imgs = (imgs*255).astype(np.uint8)
     rows = math.ceil(math.sqrt(len(imgs)))
     cols = np.ceil(len(imgs) / float(rows))
     #plt.imshow(imgs[0, :, :, :])
